postgressql dá melhor suporte a io non blocking. App gerando seu proprio UUID, precisou criar meu proprio INSERT

+++ src/main/kotlin/com/example/websocket/repository/PersonRepository.kt
@@ -3,5 +3,5 @@ package com.example.websocket.repository
 import com.example.websocket.model.Person
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
-interface PersonRepository: ReactiveCrudRepository<Person, Int> {
+interface PersonRepository: ReactiveCrudRepository<Person, Long> {
 }
